+++ src/lib/types.ts
@@ -5,6 +5,7 @@ import {PuzzleJs} from "./puzzle";
 declare global {
   interface Window {
     PuzzleJs: PuzzleJs;
+    PerformanceObserver?: any;
   }
 }