+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+# [3.5.0] - 25-07-2018
+### Fixed
+- Fragment loaded for static and waited fragments
+### Added
+- Analytics added into PuzzleLib
+
 # [3.4.8] - 19-07-2018
 ### Changed
 - Assets are being loaded in parallel